requirements-dev.txt
pytest==7.4.3
pytest-cov==4.1.0
paho-mqtt==1.6.1
flake8==6.1.0
pylint==3.0.3